backend/requirements.txt
flask
flask-cors
requests
ffmpeg-python
transformers
torch